Moved tracing span wrapper to ConcreteFutureImpl (still not working yet)
