+++ cpp/src/arrow/util/future.cc
@@ -27,6 +27,7 @@
 #include "arrow/util/checked_cast.h"
 #include "arrow/util/logging.h"
 #include "arrow/util/thread_pool.h"
+#include "arrow/util/tracing_internal.h"
 
 namespace arrow {
 
@@ -241,8 +242,23 @@ class ConcreteFutureImpl : public FutureImpl {
 
   void AddCallback(Callback callback, CallbackOptions opts) {
     CheckOptions(opts);
+    struct Wrapstruct {
+        void operator()() {
+        bool trace_valid = activeSpan->GetContext().IsValid();
+        auto scope = ::arrow::internal::tracing::GetTracer()->WithActiveSpan(activeSpan);
+        std::move(func)(*self);
+      }
+      Callback func;
+      std::shared_ptr<FutureImpl> self;
+      opentelemetry::nostd::shared_ptr<opentelemetry::trace::Span> activeSpan;
+    };
+    Wrapstruct wrapper;
+    wrapper.func = std::forward<Callback>(callback);
+    wrapper.self = shared_from_this();
+    wrapper.activeSpan = ::arrow::internal::tracing::GetTracer()->GetCurrentSpan();
+
     std::unique_lock<std::mutex> lock(mutex_);
-    CallbackRecord callback_record{std::move(callback), opts};
+    CallbackRecord callback_record{std::move(std::forward<Callback>(wrapper)), opts};
     if (IsFutureFinished(state_)) {
       lock.unlock();
       RunOrScheduleCallback(shared_from_this(), std::move(callback_record),
